mnh-clk: remove ipu reset from ipu_clock_gating function
The ipu driver does not support reset done outside of the driver. Also,
the majority, if not all, of the power saving comes from disabling ipu
clock and not from the reset.

Also, fixed the signal level for the ipu_clock_gating sysfs file since
ipu uses a clock enable bit, and not a clock gating bit.

Bug: 37857412
Change-Id: I15efeb4dcf8476ef555876ae06001daeead81c3a
Signed-off-by: Trevor Bunker <trevorbunker@google.com>