arm64: dts: mnh: enable pulldown on GPIOI9
GPIO9 is used to trigger Linux suspend from the AP. However, on our
platform, this GPIO is not connected. On at least one device (and
probably more but we hadn't discovered it yet), this GPIO is bouncing
because of noise and causing inadvertent kernel suspends. Enable the
pulldown on GPIO9 to keep it from bouncing.

Bug: 63743854
Change-Id: I7e24296e6468bb4d3b510971082e75e6c27c805b
Signed-off-by: Trevor Bunker <trevorbunker@google.com>