misc: paintbox: dma stop cleanup
1.  Make stop completion non-interruptible.
The IPU driver was using the interruptible version of the wait for
completion call but was not handling signal interrupt properly.  The
stop operation should take at most 500us so the delay to signal
processing should a signal come in is not onerous.

2.  Report cancel on both active transfers
Previously the second transfer enqueued in hardware was reported as
canceled as part of the reset operation.  This makes it difficult to
detect the case where a reset is occurring while there is an active
transfer.  This change explicitly reports the transfer as canceled
before issuing the DMA channel reset.

Change-Id: I0ae8db6083f68a24a8df7b707bc8d8de2ed6fe7b
Signed-off-by: Adam Hampson <ahampson@google.com>