misc: paintbox: enable/disable mipi on DMA availability
On a MIPI SOF, check to make sure there are sufficient buffers for
the current frame and the next.  If there is not a sufficient numbers
of buffers for the next frame then disable the stream.

If a DMA buffer that has a MIPI source is enqueued and the
corresponding MIPI stream is not enabled then enable it after the
DMA transfer has been started.

Bug: 68045211
Change-Id: Id60140c0ba0847ada0245c1bd4cc11202924bc5c
Signed-off-by: Adam Hampson <ahampson@google.com>