misc: paintbox: simplify stream enable or DMA enqueue
The previous implementation tried to reuse an existing stream
enable interface which created a race condition with regard to the
free_running and frame_count fields in the stream data structure.
This change simplifies the code and explicitly configures the stream
for free running mode and for disable on error.

Bug: 68045211

Change-Id: I9bd98c29f991de2f8a35d6d1ab7166b4668d1f58
Signed-off-by: Adam Hampson <ahampson@google.com>