misc: paintbox: do not add IRQ event if already triggered
If the IRQ entry in the wait object has already been triggered then do
not add the event again.  The event will be left in the IRQ object's
pending queue and will be reported in a later wait for interrupt.

Also do not add the IRQ event if the wait object has met its exit
criteria and the waiting thread has been signaled.

Bug: 68724906
Change-Id: I7b40f0b59e0fddbdd224cdfa1da62a22d7cbdf38
Signed-off-by: Adam Hampson <ahampson@google.com>